+++ src/features/index.js
@@ -3,8 +3,8 @@ import { combineReducers } from 'redux';
 import loadingReducer from './loadingSlice';
 import errorReducer from './errorSlice';
 import pageReducer from './pageSlice';
-import imagesReducer from './images/imagesSlice';
-import statsReducer from './images/statsSlice';
+import imagesReducer from './imagesSlice';
+import statsReducer from './statsSlice';
 
 const rootReducer = combineReducers({
     isLoading: loadingReducer,
